support return aggregation results directly

+++ src/main/java/amordleq/elasticsearch/elastic/maptile/Location.java
@@ -4,6 +4,6 @@ import lombok.Data;
 
 @Data
 public class Location {
-    Float lat;
-    Float lon;
+    Double lat;
+    Double lon;
 }
